Add a solid block texture (so we can see the world again) and align the grid background with the tile map. Update the Sprite class so it can have it's position set by using top or left coordinates instead of just the middle. Also allow sprites to have their texture scale set.
Finally, deletes the old GridPlane class since it's work can now be done via the Sprite class.

+++ SpaceWar/src/com/zeddic/common/opengl/Sprite.java
@@ -9,7 +9,7 @@ import android.graphics.Bitmap;
  */
 public class Sprite extends SimplePlane {
 
-  public Sprite(int width, int height, int resource) {
+  public Sprite(float width, float height, int resource) {
     super(width, height);
     setTexture(new Texture(resource));
   }
